+++ .qgis-plugin-ci
@@ -1,3 +1,3 @@
-plugin_path: enhanced_relation_editor_widget
+plugin_path: linking_relation_editor
 github_organization_slug: opengisch
-project_slug: qgis-enhanced-relation-editor-widget
+project_slug: qgis-linking-relation-editor
