+++ linking_relation_editor/core/plugin_helper.py
