+++ linking_relation_editor/gui/linking_relation_editor_widget_factory.py
@@ -1,7 +1,7 @@
 # -*- coding: utf-8 -*-
 # -----------------------------------------------------------
 #
-# QGIS Enhanced Relation Editor Widget Plugin
+# QGIS Linking Relation Editor Plugin
 # Copyright (C) 2021 Damiano Lombardi
 #
 # licensed under the terms of GNU GPL 2
@@ -9,20 +9,20 @@
 # -----------------------------------------------------------
 
 from qgis.gui import QgsAbstractRelationEditorWidgetFactory, QgsRelationEditorConfigWidget
-from enhanced_relation_editor_widget.gui.enhanced_relation_editor_widget import EnhancedRelationEditorWidget
+from linking_relation_editor.gui.linking_relation_editor_widget import LinkingRelationEditorWidget
 
-WIDGET_TYPE = "enhanced_relation_editor_widget"
+WIDGET_TYPE = "linking_relation_editor"
 
 
-class EnhancedRelationEditorWidgetFactory(QgsAbstractRelationEditorWidgetFactory):
+class LinkingRelationEditorWidgetFactory(QgsAbstractRelationEditorWidgetFactory):
     def type(self):
         return WIDGET_TYPE
 
     def name(self):
-        return "Enhanced relation editor widget"
+        return "Linking relation editor widget"
 
     def create(self, config, parent):
-        return EnhancedRelationEditorWidget(config, parent)
+        return LinkingRelationEditorWidget(config, parent)
 
     def configWidget(self, relation, parent):
         return QgsRelationEditorConfigWidget(relation, parent)
