+++ linking_relation_editor/gui/map_tool_select_rectangle.py
