+++ linking_relation_editor/gui/vector_layer_selection_manager.py
