+++ linking_relation_editor/tests/test_linking_relation_editor_widget_factory.py
@@ -1,0 +1,12 @@
+
+from qgis.testing import unittest, start_app
+from linking_relation_editor.gui.linking_relation_editor_widget_factory import LinkingRelationEditorWidgetFactory
+
+start_app()
+
+
+class TestLinkingRelationEditorWidgetFactory(unittest.TestCase):
+
+    def test_Instantiate(self):
+        LinkingRelationEditorWidgetFactory()
+
